refactor(OrderValidator): Order, OrderTable 양방향 의존성 제거를 위한 DIP 적용
