+++ src/main/java/kitchenpos/order/validator/OrderValidator.java
@@ -1,40 +1,11 @@
 package kitchenpos.order.validator;
 
-import kitchenpos.common.ErrorMessage;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.repository.OrderTableRepository;
-import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-@Component
-public class OrderValidator {
-
-    private final OrderTableRepository orderTableRepository;
-
-    public OrderValidator(final OrderTableRepository orderTableRepository) {
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    public void validate(OrderRequest orderRequest) {
-        OrderTable orderTable = findOrderTableById(orderRequest.getOrderTableId());
-        validateOrderTable(orderTable);
-    }
-
-    private OrderTable findOrderTableById(Long id) {
-        return orderTableRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.INVALID_ORDER_TABLE_INFO.getMessage()));
-    }
-
-    private void validateOrderTable(OrderTable orderTable) {
-        if(orderTable.isEmpty()) {
-            throw new IllegalArgumentException(ErrorMessage.EMPTY_ORDER_TABLE.getMessage());
-        }
-    }
-
-    public void validateOnGoingOrderStatus(List<Order> orders) {
-        orders.forEach(Order::validateOnGoingOrderStatus);
-    }
+public interface OrderValidator {
+    void validate(OrderRequest orderRequest);
+    void validateOnGoingOrderStatus(List<Order> orders);
 }
