+++ src/main/java/kitchenpos/ordertable/validator/OrderValidatorImpl.java
@@ -1,0 +1,43 @@
+package kitchenpos.ordertable.validator;
+
+import kitchenpos.common.ErrorMessage;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.dto.OrderRequest;
+import kitchenpos.order.validator.OrderValidator;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.repository.OrderTableRepository;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+public class OrderValidatorImpl implements OrderValidator {
+
+    private final OrderTableRepository orderTableRepository;
+
+    public OrderValidatorImpl(final OrderTableRepository orderTableRepository) {
+        this.orderTableRepository = orderTableRepository;
+    }
+
+    @Override
+    public void validate(OrderRequest orderRequest) {
+        OrderTable orderTable = findOrderTableById(orderRequest.getOrderTableId());
+        validateOrderTable(orderTable);
+    }
+
+    private OrderTable findOrderTableById(Long id) {
+        return orderTableRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.INVALID_ORDER_TABLE_INFO.getMessage()));
+    }
+
+    private void validateOrderTable(OrderTable orderTable) {
+        if(orderTable.isEmpty()) {
+            throw new IllegalArgumentException(ErrorMessage.EMPTY_ORDER_TABLE.getMessage());
+        }
+    }
+
+    @Override
+    public void validateOnGoingOrderStatus(List<Order> orders) {
+        orders.forEach(Order::validateOnGoingOrderStatus);
+    }
+}
